refactor(inspection-service): extract inspection-publish sub-service
Move publishInspection (status flip + automation + auto-sign),
computePublishReadiness, getReportGate, and getInspectionHub into
InspectionPublishService. getInspectionHub composes getPeopleCard (core) via
the facade and computePublishReadiness internally; the publishInspection
automation chain (update -> fireAutomation -> auto-sign) stays intact as a
byte-identical move. The facade delegates (full original signatures preserved).
