+++ server/services/inspection.service.ts
@@ -1,6 +1,6 @@
 import { drizzle } from 'drizzle-orm/d1';
 import { eq, and, or, lt, gte, lte, sql, inArray, desc } from 'drizzle-orm';
-import { inspections, inspectionResults, templates, users, services, inspectionServices, tenantConfigs, invoices, tenants, agreementRequests, agreements, reportVersions } from '../lib/db/schema';
+import { inspections, inspectionResults, templates, users, services, inspectionServices, tenantConfigs, agreementRequests, reportVersions } from '../lib/db/schema';
 import { contacts } from '../lib/db/schema/contact';
 import { Errors } from '../lib/errors';
 import { getRatingBucket, type RatingLevel } from '../lib/report-utils';
@@ -17,8 +17,6 @@ import type { AgreementService } from './agreement.service';
 import { findingKey, DEFAULT_UNIT } from '../lib/finding-key';
 import { parseReinspectionStatuses, isOpenStatus } from '../lib/reinspection-status';
 import { type ReportMediaContext } from '../lib/report-video';
-import { InvoiceService } from './invoice.service';
-import type { TemplateSchemaV2 } from '../types/template-schema';
 import { type ImagesBinding } from '../lib/media/strip-exif';
 import { type PdfSettings } from '../lib/pdf-settings';
 import { INSPECTION_STATUS } from '../lib/status/inspection-status';
@@ -54,6 +52,7 @@ import { InspectionAnnotationsService } from './inspection/inspection-annotation
 import { InspectionPhotoService } from './inspection/inspection-photo.service';
 import { InspectionResultsService } from './inspection/inspection-results.service';
 import { InspectionReportService } from './inspection/inspection-report.service';
+import { InspectionPublishService } from './inspection/inspection-publish.service';
 export {
     resolveCoverUrl,
     sanitizeDefectStates,
@@ -86,6 +85,7 @@ export class InspectionService {
     private readonly photo: InspectionPhotoService;
     private readonly results: InspectionResultsService;
     private readonly report: InspectionReportService;
+    private readonly publish: InspectionPublishService;
 
     constructor(private db: D1Database, r2?: R2Bucket, private sdb?: ScopedDB, kv?: KVNamespace, images?: ImagesBinding) {
         this.sharing = new InspectionSharingService(db, r2, sdb, kv, images);
@@ -95,6 +95,7 @@ export class InspectionService {
         this.photo = new InspectionPhotoService(db, r2, sdb, kv, images, this);
         this.results = new InspectionResultsService(db, r2, sdb, kv, images);
         this.report = new InspectionReportService(db, r2, sdb, kv, images);
+        this.publish = new InspectionPublishService(db, r2, sdb, kv, images, this);
     }
 
     private getDrizzle() {
@@ -1074,123 +1075,7 @@ export class InspectionService {
         amountCents: number | null;
         currency: string | null;
     } | null> {
-        const db = this.getDrizzle();
-        const insp = await db.select({
-            id:                inspections.id,
-            propertyAddress:   inspections.propertyAddress,
-            date:              inspections.date,
-            inspectorId:       inspections.inspectorId,
-            paymentRequired:   inspections.paymentRequired,
-            paymentStatus:     inspections.paymentStatus,
-            agreementRequired: inspections.agreementRequired,
-        }).from(inspections)
-            .where(and(eq(inspections.id, inspectionId), eq(inspections.tenantId, tenantId)))
-            .get();
-        if (!insp) return null;
-
-        // Resolve the outstanding gate. Agreement before payment (signed first).
-        let reason: 'payment' | 'agreement' | null = null;
-        let agreementToken: string | null = null;
-        if (insp.agreementRequired === true) {
-            const signed = await db.select({ id: agreementRequests.id })
-                .from(agreementRequests)
-                .where(and(
-                    eq(agreementRequests.inspectionId, inspectionId),
-                    eq(agreementRequests.tenantId, tenantId),
-                    eq(agreementRequests.status, 'signed'),
-                ))
-                .limit(1);
-            if (signed.length === 0) {
-                reason = 'agreement';
-                const pending = await db.select({ token: agreementRequests.token })
-                    .from(agreementRequests)
-                    .where(and(
-                        eq(agreementRequests.inspectionId, inspectionId),
-                        eq(agreementRequests.tenantId, tenantId),
-                    ))
-                    .orderBy(desc(agreementRequests.createdAt))
-                    .limit(1)
-                    .get();
-                agreementToken = pending?.token ?? null;
-            }
-        }
-        // Payment-outstanding is computed independently of `reason` so the
-        // dual-gate (agreement AND payment) case can route to combined checkout.
-        const paymentOutstanding = insp.paymentRequired === true && insp.paymentStatus !== 'paid';
-        if (!reason && paymentOutstanding) {
-            reason = 'payment';
-        }
-        if (!reason) return null;   // not gated — nothing to surface
-
-        // Track I-a Task 7 — both gates outstanding → combined "Sign & pay".
-        const bothOutstanding = reason === 'agreement' && paymentOutstanding;
-
-        const branding = await db.select({ siteName: tenantConfigs.siteName, primaryColor: tenantConfigs.primaryColor })
-            .from(tenantConfigs).where(eq(tenantConfigs.tenantId, tenantId)).get();
-
-        let inspector: { name: string | null; email: string | null; phone: string | null; licenseNumber: string | null } | undefined;
-        if (insp.inspectorId) {
-            inspector = await db.select({
-                name: users.name, email: users.email, phone: users.phone, licenseNumber: users.licenseNumber,
-            }).from(users)
-                .where(and(eq(users.id, insp.inspectorId), eq(users.tenantId, tenantId)))
-                .get();
-        }
-
-        // Surface the invoice amount whenever payment is part of the gate (the
-        // payment-only page AND the combined Sign & pay page both show it).
-        let amountCents: number | null = null;
-        if (paymentOutstanding) {
-            const invoice = await db.select({ amountCents: invoices.amountCents })
-                .from(invoices)
-                .where(and(eq(invoices.tenantId, tenantId), eq(invoices.inspectionId, inspectionId)))
-                .orderBy(desc(invoices.createdAt))
-                .limit(1)
-                .get();
-            amountCents = invoice?.amountCents ?? null;
-        }
-
-        // Reconstruct the first outstanding signer's tier-2 link token
-        // server-side. Used by BOTH the combined "Sign & pay" checkout URL and
-        // the agreement-only sign URL — `agreementRequests.token` is an
-        // UNDISTRIBUTED placeholder for envelope-v2 (real tokens live per-signer),
-        // so routing the customer to it would 404. When the helper is unavailable
-        // or yields no outstanding signer, fall back to the legacy envelope token
-        // (still resolves for legacy `createSigningRequest` envelopes whose
-        // plaintext token IS distributed) — last resort, never break those.
-        let signerLink: string | null = null;
-        if ((bothOutstanding || reason === 'agreement') && agreementService) {
-            signerLink = await agreementService.getFirstOutstandingSignerLink(tenantId, inspectionId);
-        }
-        const agreementLinkToken = signerLink ?? agreementToken;
-
-        const actionUrl = bothOutstanding && signerLink
-            ? `/checkout/${tenantSlug}/${signerLink}`
-            : reason === 'payment'
-                ? `/invoice/${inspectionId}`
-                : (agreementLinkToken ? `/agreements/sign/${tenantSlug}/${agreementLinkToken}` : `/report-gate/${tenantSlug}/${inspectionId}`);
-
-        const actionLabel = bothOutstanding && signerLink
-            ? 'Sign & pay'
-            : reason === 'payment' ? 'Pay invoice' : 'Sign agreement';
-
-        return {
-            reason,
-            companyName: branding?.siteName ?? 'OpenInspection',
-            // A-10 — nullable: null means "tenant set no accent", the page
-            // keeps the platform design tokens (no per-surface fallback hex).
-            primaryColor: branding?.primaryColor ?? null,
-            actionUrl,
-            actionLabel,
-            propertyAddress: insp.propertyAddress ?? null,
-            inspectorName: inspector?.name ?? null,
-            inspectorEmail: inspector?.email ?? null,
-            inspectorPhone: inspector?.phone ?? null,
-            inspectorLicense: inspector?.licenseNumber ?? null,
-            scheduledDate: insp.date ?? null,
-            amountCents,
-            currency: amountCents != null ? 'USD' : null,
-        };
+        return this.publish.getReportGate(inspectionId, tenantId, tenantSlug, agreementService);
     }
 
     /**
@@ -1245,113 +1130,7 @@ export class InspectionService {
         invoice: { id: string; status: string; amountCents: number; sentAt: string | null; paidAt: string | null } | null;
         publishReadiness: { ready: boolean; blockingCount: number };
     } | null> {
-        const db = this.getDrizzle();
-
-        // Authority row — gate on existence + tenant ownership first.
-        const insp = await db.select().from(inspections)
-            .where(and(eq(inspections.id, inspectionId), eq(inspections.tenantId, tenantId)))
-            .get();
-        if (!insp) return null;
-
-        // Service lines — effective price = priceOverride ?? priceSnapshot
-        // (P-4 authority chain, tier 2). Tenant-scoped on both columns.
-        const serviceRows = await db.select({
-            id:            inspectionServices.id,
-            nameSnapshot:  inspectionServices.nameSnapshot,
-            priceSnapshot: inspectionServices.priceSnapshot,
-            priceOverride: inspectionServices.priceOverride,
-        }).from(inspectionServices)
-            .where(and(
-                eq(inspectionServices.tenantId, tenantId),
-                eq(inspectionServices.inspectionId, inspectionId),
-            ))
-            .all();
-
-        // Tenant's agreement templates — drives a "send agreement" dropdown later.
-        const agreementRows = await db.select({ id: agreements.id, name: agreements.name })
-            .from(agreements)
-            .where(eq(agreements.tenantId, tenantId))
-            .orderBy(desc(agreements.createdAt))
-            .all();
-
-        // Agreement requests for this inspection, newest first.
-        const requestRows = await db.select({
-            id:          agreementRequests.id,
-            status:      agreementRequests.status,
-            clientEmail: agreementRequests.clientEmail,
-            signedAt:    agreementRequests.signedAt,
-            createdAt:   agreementRequests.createdAt,
-        }).from(agreementRequests)
-            .where(and(
-                eq(agreementRequests.tenantId, tenantId),
-                eq(agreementRequests.inspectionId, inspectionId),
-            ))
-            .orderBy(desc(agreementRequests.createdAt))
-            .all();
-
-        // Reused primitives. getPeopleCard/computePublishReadiness throw NotFound
-        // when the row is absent — but we already confirmed it exists above, so
-        // they resolve. InvoiceService is constructed inline (it takes only a
-        // D1Database, same handle this service holds) per the DI guidance: no
-        // constructor-chain redesign, just compose the read.
-        const invoiceSvc = new InvoiceService(this.db);
-        const [people, readiness, invoice] = await Promise.all([
-            this.getPeopleCard(inspectionId, tenantId),
-            this.computePublishReadiness(inspectionId, tenantId),
-            invoiceSvc.findByInspectionId(tenantId, inspectionId),
-        ]);
-
-        return {
-            inspection: {
-                id:                insp.id,
-                propertyAddress:   insp.propertyAddress,
-                clientName:        insp.clientName ?? null,
-                clientEmail:       insp.clientEmail ?? null,
-                clientPhone:       insp.clientPhone ?? null,
-                clientContactId:   insp.clientContactId ?? null,
-                status:            insp.status,
-                reportStatus:      insp.reportStatus as string,
-                date:              insp.date ?? null,
-                inspectorId:       insp.inspectorId ?? null,
-                templateId:        insp.templateId ?? null,
-                price:             insp.price,
-                paymentStatus:     insp.paymentStatus,
-                paymentRequired:   insp.paymentRequired === true,
-                agreementRequired: insp.agreementRequired === true,
-                coverPhoto:        insp.coverPhotoId ?? null,
-                referredByAgentId: insp.referredByAgentId ?? null,
-                sellingAgentId:    insp.sellingAgentId ?? null,
-                createdAt:         safeISODate(insp.createdAt),
-            },
-            tenantSlug,
-            people,
-            services: serviceRows.map(s => ({
-                id:        s.id,
-                name:      s.nameSnapshot,
-                priceCents: s.priceOverride ?? s.priceSnapshot,
-            })),
-            agreements: agreementRows.map(a => ({ id: a.id, name: a.name })),
-            agreementRequests: requestRows.map(r => ({
-                id:          r.id,
-                status:      r.status,
-                clientEmail: r.clientEmail,
-                signedAt:    r.signedAt ? safeISODate(r.signedAt) : null,
-                createdAt:   safeISODate(r.createdAt),
-            })),
-            invoice: invoice
-                ? {
-                    id:         invoice.id,
-                    status:     invoice.status,
-                    amountCents: invoice.amountCents,
-                    sentAt:     invoice.sentAt,
-                    paidAt:     invoice.paidAt,
-                }
-                : null,
-            publishReadiness: {
-                ready:         readiness.ready,
-                blockingCount: readiness.blockingDefects.length,
-            },
-        };
+        return this.publish.getInspectionHub(inspectionId, tenantId, tenantSlug);
     }
 
     /**
@@ -1569,67 +1348,7 @@ export class InspectionService {
         recipients?: Array<{ contactId: string | null; channels: Array<'email' | 'text'> }>;
         sendAgreementCopy?: boolean;
     }) {
-        const db = this.getDrizzle();
-
-        const inspection = await db.select().from(inspections)
-            .where(and(eq(inspections.id, inspectionId), eq(inspections.tenantId, tenantId)))
-            .get();
-        if (!inspection) throw Errors.NotFound('Inspection not found');
-        if (inspection.status !== INSPECTION_STATUS.COMPLETED) throw Errors.BadRequest('Inspection must be completed before publishing the report.');
-
-        await db.update(inspections)
-            .set({ reportStatus: REPORT_STATUS.PUBLISHED })
-            .where(and(eq(inspections.id, inspectionId), eq(inspections.tenantId, tenantId)));
-        // Await so AutomationService.trigger actually inserts automation_logs
-        // before the response goes out — the prior fire-and-forget pattern
-        // dangled the promise so CF terminated the isolate before the insert
-        // completed (and ditto for inspection.confirmed / cancelled / created
-        // below — all four paths now block on trigger).
-        await fireAutomation(this.db, tenantId, inspectionId, 'report.published');
-
-        // Spec 5H D2 — auto-sign on publish: if the inspection has the flag
-        // enabled AND the assigned inspector has a saved signature, inject
-        // _inspector_signature into inspection_results.data so the published
-        // report renders with the signature without requiring a manual step.
-        const inspForSign = await db.select().from(inspections)
-            .where(and(eq(inspections.id, inspectionId), eq(inspections.tenantId, tenantId)))
-            .get();
-        if (inspForSign?.autoSignOnPublish && inspForSign.inspectorId) {
-            const inspector = await db.select().from(users)
-                .where(eq(users.id, inspForSign.inspectorId)).get();
-            if (inspector?.defaultSignatureBase64) {
-                const resultsRow = await db.select().from(inspectionResults)
-                    .where(eq(inspectionResults.inspectionId, inspectionId)).get();
-                const data: Record<string, unknown> = (resultsRow?.data as Record<string, unknown>) ?? {};
-                data._inspector_signature = {
-                    signatureBase64: inspector.defaultSignatureBase64,
-                    signedAt:        Date.now(),
-                    userId:          inspector.id,
-                    auto:            true,
-                };
-                if (resultsRow) {
-                    await db.update(inspectionResults)
-                        .set({ data: data as object, lastSyncedAt: new Date() })
-                        .where(eq(inspectionResults.id, resultsRow.id));
-                } else {
-                    await db.insert(inspectionResults).values({
-                        id:           crypto.randomUUID(),
-                        tenantId,
-                        inspectionId,
-                        data:         data as object,
-                        lastSyncedAt: new Date(),
-                    });
-                }
-            }
-        }
-
-        const tenantRow = await db.select({ slug: tenants.slug })
-            .from(tenants).where(eq(tenants.id, tenantId)).get();
-        const tenantSlug = tenantRow?.slug ?? '';
-        return {
-            reportUrl: `/report/${tenantSlug}/${inspectionId}`,
-            reportStatus: REPORT_STATUS.PUBLISHED,
-        };
+        return this.publish.publishInspection(inspectionId, tenantId, _options);
     }
 
     async confirmInspection(tenantId: string, id: string): Promise<void> {
@@ -1717,60 +1436,7 @@ export class InspectionService {
      * takes precedence over the live template.schema.
      */
     async computePublishReadiness(inspectionId: string, tenantId: string): Promise<PublishReadiness> {
-        const db = this.getDrizzle();
-
-        const inspection = await db.select().from(inspections)
-            .where(and(eq(inspections.id, inspectionId), eq(inspections.tenantId, tenantId)))
-            .get();
-        if (!inspection) throw Errors.NotFound('Inspection not found');
-
-        const template = inspection.templateId
-            ? await db.select().from(templates)
-                .where(and(eq(templates.id, inspection.templateId as string), eq(templates.tenantId, tenantId)))
-                .get()
-            : null;
-
-        const resultsRow = await db.select().from(inspectionResults)
-            .where(and(eq(inspectionResults.inspectionId, inspectionId), eq(inspectionResults.tenantId, tenantId)))
-            .get();
-
-        // Prefer per-inspection snapshot over live template schema (mirrors getReportData).
-        const inspectionSnapshotRaw = (inspection as unknown as { templateSnapshot?: unknown }).templateSnapshot;
-        const inspectionSnapshot = inspectionSnapshotRaw
-            ? (typeof inspectionSnapshotRaw === 'string' ? JSON.parse(inspectionSnapshotRaw as string) : inspectionSnapshotRaw)
-            : null;
-        const hasInspectionSnapshot = inspectionSnapshot
-            && typeof inspectionSnapshot === 'object'
-            && Array.isArray((inspectionSnapshot as { sections?: unknown }).sections)
-            && (inspectionSnapshot as { sections: unknown[] }).sections.length > 0;
-
-        const rawSchema = hasInspectionSnapshot
-            ? inspectionSnapshot
-            : template?.schema
-                ? (typeof template.schema === 'string' ? JSON.parse(template.schema) : template.schema)
-                : { sections: [] };
-
-        interface RawSchemaData { sections?: unknown[] }
-        const schemaData: TemplateSchemaV2 = Array.isArray(rawSchema)
-            ? ({ schemaVersion: 2, sections: [{ id: 'general', title: 'General', items: rawSchema }] } as unknown as TemplateSchemaV2)
-            : (rawSchema as RawSchemaData).sections
-                ? rawSchema as TemplateSchemaV2
-                : ({ schemaVersion: 2, sections: [] } as unknown as TemplateSchemaV2);
-
-        const resultData: Record<string, unknown> = resultsRow?.data
-            ? (typeof resultsRow.data === 'string' ? JSON.parse(resultsRow.data) : resultsRow.data) as Record<string, unknown>
-            : {};
-
-        // Track H (IA-7 / P-6②) — effective requirement: per-inspection
-        // override beats the tenant default; both unset → 'none' (loose).
-        const cfgRow = await db.select({ requireDefectFields: tenantConfigs.requireDefectFields })
-            .from(tenantConfigs)
-            .where(eq(tenantConfigs.tenantId, tenantId))
-            .get();
-        const override = (inspection as unknown as { requireDefectFieldsOverride?: RequireDefectFields | null }).requireDefectFieldsOverride;
-        const requirement = resolveRequireDefectFields(override, cfgRow?.requireDefectFields);
-
-        return computePublishReadinessFromState(schemaData, resultData, requirement);
+        return this.publish.computePublishReadiness(inspectionId, tenantId);
     }
 
     /**
